refactor: Extract view for task list
