+++ lib/src/digit_span_tasks/digit_span_tasks_view.dart
@@ -1,0 +1,65 @@
+import 'package:digit_span_tasks/digit_span_tasks.dart';
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({super.key});
+
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            ElevatedButton(
+              onPressed: runDigitSpanForward,
+              child: Text(
+                'Digit Span Forward',
+                style: Theme.of(context).textTheme.titleLarge,
+              ),
+            ),
+            const SizedBox(height: 10),
+            ElevatedButton(
+              onPressed: runDigitSpanBackwards,
+              child: Text(
+                'Digit Span Backwards',
+                style: Theme.of(context).textTheme.titleLarge,
+              ),
+            ),
+          ],
+        ),
+      ),
+      backgroundColor: Colors.grey[300],
+    );
+  }
+
+  void runDigitSpanForward() async {
+    UserConfig config = UserConfig(
+      stimListPractice: ['01', '234'],
+      stimListExperimental: ['5678', '01567', '987654'],
+    );
+    DigitSpanTasksData data = await Get.to(() => DigitSpanForward(
+          config: config,
+        ));
+    // ignore: avoid_print
+    print('\n\n\nFORWARD data \n $data');
+  }
+
+  void runDigitSpanBackwards() async {
+    UserConfig config = UserConfig(
+      stimListPractice: ['23', '567'],
+      stimListExperimental: ['0123', '45678', '901234'],
+    );
+    DigitSpanTasksData data = await Get.to(() => DigitSpanBackwards(
+          config: config,
+        ));
+    // ignore: avoid_print
+    print('\n\n\nBAKWARDS data \n $data');
+  }
+}
